Support float type for size and weight attributes of Item

+++ src/mmobot/db/models/item.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column
-from sqlalchemy import Integer
+from sqlalchemy import Float
 from sqlalchemy import String
 
 from .base import Base
@@ -10,8 +10,8 @@ class Item(Base):
 
     id = Column(String(40), primary_key=True)
     item_type = Column(String(20))
-    size = Column(Integer)
-    weight = Column(Integer)
+    size = Column(Float)
+    weight = Column(Float)
 
     __mapper_args__ = {
         'polymorphic_identity': 'item',
